flask_app/requirements.txt
Flask>=2.0
python-dotenv>=0.19
docker>=5.0
Flask-CORS>=3.0
